+++ src/components/Checkout/DonationItem.vue
@@ -73,7 +73,7 @@
 							</button>
 						</div>
 
-						<div v-if="hasLoans">
+						<div>
 							<div
 								class="donation-tagline tw-text-small tw-text-secondary tw-my-1"
 								data-testid="basket-donation-tagline"
@@ -189,7 +189,7 @@
 								class="tw-flex-1 md:tw-flex-grow tw-text-h3"
 								data-testid="basket-donation-title"
 							>
-								Help our team find more borrowers
+								{{ basketDonateModulesExpDonationHeader }}
 							</h2>
 							<button
 								class="donation-amount md:tw-hidden tw-flex-none tw-align-middle"
@@ -241,16 +241,12 @@
 							</button>
 						</div>
 
-						<div v-if="hasLoans">
+						<div>
 							<div
 								class="donation-tagline tw-text-small tw-text-secondary tw-my-1 tw-max-w-2xl"
 								data-testid="basket-donation-tagline"
 							>
-								<p class="tw-text-primary tw-text-base">
-									<em>"Our team works one-on-one with partners so they can find and support
-										borrowers with services like training and healthcare.
-										Your tips make our work possible."</em><br>
-									&dash; Carolina, Borrower success team.
+								<p class="tw-text-primary tw-text-base" v-html="basketDonateModulesExpDonationTagline">
 								</p>
 							</div>
 							<button
@@ -347,7 +343,7 @@
 				<span class="tw-text-secondary tw-text-small tw-pr-1">Donate to Kiva</span>
 				<hr class="tw-flex-grow">
 			</div>
-			<h3>Help our team find more borrowers</h3>
+			<h3>{{ basketDonateModulesExpDonationHeader }}</h3>
 			<div class="tw-flex tw-flex-col md:tw-flex-row tw-pb-5 tw-mt-2">
 				<!-- donation image -->
 				<div class="tw-hidden md:tw-block tw-flex-none md:tw-mr-3 lg:tw-mr-4.5">
@@ -392,16 +388,12 @@
 							</button>
 						</div>
 
-						<div v-if="hasLoans">
+						<div>
 							<div
 								class="donation-tagline tw-text-small tw-text-secondary tw-mb-1 tw-max-w-2xl"
 								data-testid="basket-donation-tagline"
 							>
-								<p class="tw-text-primary tw-text-base">
-									<em>"Our team works one-on-one with partners so they can find and support
-										borrowers with services like training and healthcare.
-										Your tips make our work possible."</em><br>
-									&dash; Carolina, Borrower success team.
+								<p class="tw-text-primary tw-text-base" v-html="basketDonateModulesExpDonationTagline">
 								</p>
 							</div>
 							<button
@@ -493,7 +485,175 @@
 				</div>
 			</div>
 		</template>
+		<template v-if="donateItemExperimentVersion === 'd'">
+			<div class="tw-flex tw-flex-col md:tw-flex-row tw-pb-5">
+				<!-- donation image -->
+				<div class="tw-hidden md:tw-block tw-flex-none md:tw-mr-3 lg:tw-mr-4.5">
+					<img
+						class="donation-img tw-w-12 lg:tw-w-13 tw-h-12 lg:tw-h-13 tw-rounded"
+						:src="imageRequire(`./peace-sign-holding-money.svg`)"
+						alt="donation line item image"
+						data-testid="basket-donation-image"
+					>
+				</div>
 
+				<!-- donation text -->
+				<div class="tw-flex-auto">
+					<div class="tw-mb-0.5">
+						<div class="tw-w-full tw-flex">
+							<h2
+								class="tw-flex-1 md:tw-flex-grow tw-text-h3"
+								data-testid="basket-donation-title"
+							>
+								{{ basketDonateModulesExpDonationHeader }}
+							</h2>
+							<button
+								class="donation-amount md:tw-hidden tw-flex-none tw-align-middle"
+								data-testid="basket-donation-edit-button-mobile-pencil"
+								v-kv-track-event="['basket', 'Edit Donation']"
+								@click="enterEditDonation"
+								title="Edit Donation"
+							>
+								<kv-material-icon
+									role="img"
+									aria-label="Edit Donation"
+									title="Edit Donation"
+									class="edit-donation tw-text-action
+									tw-w-3.5
+									tw-h-3.5
+									tw-py-0.5
+									md:tw-hidden"
+									name="pencil"
+									:icon="mdiPencil"
+								/>
+							</button>
+						</div>
+
+						<div
+							v-show="!editDonation"
+							class="md:tw-hidden"
+						>
+							<button
+								class="donation-amount"
+								data-testid="basket-donation-edit-button-mobile-amount"
+								v-kv-track-event="['basket', 'Edit Donation']"
+								@click="enterEditDonation"
+								title="Edit Donation"
+							>
+								{{ formattedAmount }}
+								<kv-material-icon
+									role="img"
+									aria-label="Edit Donation"
+									title="Edit Donation"
+									class="edit-donation tw-text-action
+									tw-w-3.5
+									tw-h-3.5
+									tw-py-0.5
+									tw-hidden md:tw-inline-block
+									"
+									name="pencil"
+									:icon="mdiPencil"
+								/>
+							</button>
+						</div>
+
+						<div>
+							<div
+								class="donation-tagline tw-text-small tw-text-secondary tw-my-1 tw-max-w-2xl"
+								data-testid="basket-donation-tagline"
+							>
+								<p class="tw-text-primary tw-text-base" v-html="basketDonateModulesExpDonationTagline">
+								</p>
+							</div>
+							<button
+								class="tw-text-small tw-text-link"
+								data-testid="basket-donation-info-lightbox"
+								@click="triggerDefaultLightbox"
+								v-kv-track-event="['basket', 'Donation Info Lightbox', 'Open Lightbox']"
+							>
+								{{ donationDetailsLink }}
+							</button>
+						</div>
+					</div>
+				</div>
+
+				<!-- donation total -->
+				<div
+					class="
+				tw-flex-none
+				tw-w-full
+				md:tw-w-auto
+				md:tw-ml-3
+				lg:tw-ml-4.5
+				tw-mt-1.5
+				md:tw-mt-0"
+				>
+					<div
+						v-show="!editDonation"
+						class="tw-hidden md:tw-block tw-text-right"
+					>
+						<button
+							class="donation-amount"
+							data-testid="basket-donation-edit-button-combined"
+							v-kv-track-event="['basket', 'Edit Donation']"
+							@click="enterEditDonation"
+							title="Edit Donation"
+						>
+							{{ formattedAmount }}
+							<kv-material-icon
+								role="img"
+								aria-label="Edit Donation"
+								title="Edit Donation"
+								class="edit-donation tw-text-action
+								tw-w-3.5
+								tw-h-3.5
+								tw-py-0.5
+								tw-align-bottom
+								"
+								name="pencil"
+								:icon="mdiPencil"
+							/>
+						</button>
+					</div>
+					<div v-show="editDonation" class="small-12 columns donation-amount-input-wrapper">
+						<kv-text-input
+							class="donation-amount-input"
+							data-testid="basket-donation-edit-input"
+							name="donation"
+							id="donation"
+							v-model="amount"
+							@blur="validateInput"
+							@keyup.enter.prevent="updateDonation()"
+						/>
+						<kv-button
+							variant="secondary"
+							class="update-donation-inline-button"
+							data-testid="basket-donation-edit-submit"
+							@click="updateDonation()"
+						>
+							Update
+						</kv-button>
+						<button
+							class="show-for-medium remove-wrapper"
+							@click="updateLoanAmount('remove')"
+							data-testid="basket-donation-remove"
+						>
+							<kv-material-icon
+								class="remove-x tw-text-tertiary"
+								name="small-x"
+								:from-sprite="true"
+								title="Remove donation"
+							/>
+						</button>
+					</div>
+					<donate-repayments
+						v-if="hasLoans"
+						@updating-totals="$emit('updating-totals', $event)"
+						@refreshtotals="$emit('refreshtotals')"
+					/>
+				</div>
+			</div>
+		</template>
 		<!-- Donation nudge lightbox -->
 		<donation-nudge-lightbox
 			ref="nudgeLightbox"
@@ -688,6 +848,39 @@ export default {
 			}];
 			return this.hasCustomDonation ? basePercentageRows : lowPercentage.concat(basePercentageRows);
 		},
+		basketDonateModulesExpDonationHeader() {
+			if (!this.hasLoans) {
+				return this.donationTitle;
+			}
+
+			switch (this.donateItemExperimentVersion) {
+				// variant a is the control, and uses this.donationTitle
+				case 'b':
+				case 'c':
+				default:
+					return 'Help our team find more borrowers';
+				case 'd':
+					return 'Optional: cover the cost of your loan';
+			}
+		},
+		basketDonateModulesExpDonationTagline() {
+			if (!this.hasLoans) {
+				// eslint-disable-next-line max-len
+				return 'As an international nonprofit, we’re committed to transparency about how your funds are used. 100% of money lent on Kiva goes to supporting borrowers. Your donations help us cover the costs of distributing these loans and reach more borrowers across the globe.';
+			}
+
+			switch (this.donateItemExperimentVersion) {
+				// variant a is the control, and uses this.donationTagLine
+				case 'b':
+				case 'c':
+				default:
+					// eslint-disable-next-line max-len
+					return '<em>"Our team works one-on-one with partners so they can find and support borrowers with services like training and healthcare.	Your tips make our work possible."</em><br> &dash; Carolina, Borrower success team.';
+				case 'd':
+					// eslint-disable-next-line max-len
+					return 'Loans aren\'t free for us to manage. Our expenses include borrower due diligence, partner training, site maintenance and more. We rely on our generous donors to break even.';
+			}
+		}
 	},
 	methods: {
 		updateDonationTo(amount) {
@@ -711,15 +904,13 @@ export default {
 			this.defaultLbVisible = false;
 		},
 		setupContentfulContent() {
-			if (this.hasLoans) {
-				const contentfulContentRaw = this.apollo.readQuery({
-					query: donationDataQuery,
-				});
-				const pageEntry = contentfulContentRaw?.contentful?.entries?.items?.[0] ?? null;
-				const pageData = pageEntry ? processPageContentFlat(pageEntry) : null;
-				// eslint-disable-next-line max-len
-				this.dynamicDonationItem = pageData?.page?.contentGroups?.checkoutDonationItem?.contents?.[0]?.richText ?? '';
-			}
+			const contentfulContentRaw = this.apollo.readQuery({
+				query: donationDataQuery,
+			});
+			const pageEntry = contentfulContentRaw?.contentful?.entries?.items?.[0] ?? null;
+			const pageData = pageEntry ? processPageContentFlat(pageEntry) : null;
+			// eslint-disable-next-line max-len
+			this.dynamicDonationItem = pageData?.page?.contentGroups?.checkoutDonationItem?.contents?.[0]?.richText ?? '';
 		},
 		updateDonation() {
 			this.editDonation = false;